Make the RiveSharp NuGet package cross platform
Convert RiveSharp to a netstandard2.0 library: As a netstandard2.0
library it can be referenced by all kinds of C# projects, including UWP,
thus removing the need for a secondary "RiveSharp.UWP" project.

Build x86 and x64 versions of rive.dll and include both in the NuGet
package. This required fixing some warnings that came up when building
x86. ARM builds can follow this same pattern when we get to them.

Diffs=
e617e4448 Make the RiveSharp NuGet package cross platform

+++ src/core/binary_reader.cpp
@@ -45,14 +45,14 @@ std::string BinaryReader::readString() {
         return std::string();
     }
 
-    std::vector<char> rawValue(length + 1);
+    std::vector<char> rawValue((size_t)length + 1);
     auto readBytes = decode_string(length, m_Position, m_Bytes.end(), &rawValue[0]);
     if (readBytes != length) {
         overflow();
         return std::string();
     }
     m_Position += readBytes;
-    return std::string(rawValue.data(), length);
+    return std::string(rawValue.data(), (size_t)length);
 }
 
 Span<const uint8_t> BinaryReader::readBytes() {
